third-party: fix use after free in curl
In some cases after a request is completed the curl easy handle remains
in some non-complete state, which causes this handle to remain its
connection owner. The easy handle is cleaned up in a callback, but as
the connection handle doesn't know that this easy handle was destroyed
it continues to operate with one.